requirements.txt
httpx==0.28.1
python-dotenv==1.0.1
aiogram==3.18.0